city map working

+++ simulation/de.pgalise.simulation.shared/src/main/java/de/pgalise/simulation/shared/event/weather/WeatherEventType.java
@@ -13,7 +13,7 @@ import de.pgalise.simulation.shared.event.EventType;
  * @author richter
  */
 public interface WeatherEventType extends EventType {
-	int getMinValue();
+	float getMinValue();
 	
-	int getMaxValue();
+	float getMaxValue();
 }
